Minor code move around
+ make `kernels` module `pub(crate)`
+ remove `kernels` test in favor of adding it in the
  `robust` crate
+ replace existing `Winding` trait with robust variant
+ fix `Orient` trait to use new `Winding` trait
