+++ geo/src/algorithm/mod.rs
@@ -1,5 +1,5 @@
 /// Kernels to compute various predicates
-pub mod kernels;
+pub(crate) mod kernels;
 
 /// Calculate the area of the surface of a `Geometry`.
 pub mod area;
@@ -63,8 +63,6 @@ pub mod vincenty_distance;
 pub mod vincenty_length;
 /// Calculate and work with the winding order of `Linestring`s.
 pub mod winding_order;
-/// Calculate the winding order of `LineString`s using robust predicates.
-pub mod robust_winding_order;
 /// Locate a point along a `Line` or `LineString`.
 pub mod line_locate_point;
 /// Interpolate a point along a `Line` or `LineString`.
